requirements-simple.txt
Flask==3.1.1
flask-cors==6.0.1
openai==1.12.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
requests==2.31.0
Werkzeug==3.1.3
python-dotenv==1.0.0 